backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
websockets==13.0
pydantic==2.10.0
sqlalchemy==2.0.25
python-dotenv==1.0.1
